refactor(*): change path to index.module.css
change path to index.module.css to shorter one
