+++ search-films/src/components/MovieCard/index.module.css
